+++ apps/api/controllers/kigo/legacy/booking.rb
@@ -1,0 +1,20 @@
+require_relative "../booking"
+
+module API::Controllers::Kigo::Legacy
+
+  # API::Controllers::Kigo::Legacy::Booking
+  #
+  # Performs create booking for properties from Kigo::Legacy.
+  class Booking
+    include API::Controllers::Booking
+
+    def create_booking(params)
+      credentials = Concierge::Credentials.for("KigoLegacy")
+      Kigo::Legacy.new(credentials).book(params)
+    end
+
+    def supplier_name
+      Kigo::Legacy::SUPPLIER_NAME
+    end
+  end
+end